[executors] separate bg exec impl from api
