+++ libraries/executors/executors-mcin-20100128-2304-mcb1.9-pre2/src/main/java/net/ornithemc/osl/executors/impl/mixin/client/MinecraftMixin.java
@@ -12,8 +12,8 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
 import net.minecraft.client.Minecraft;
 
-import net.ornithemc.osl.executors.api.Executors;
 import net.ornithemc.osl.executors.api.MainThreadExecutor;
+import net.ornithemc.osl.executors.impl.Executors;
 
 @Mixin(Minecraft.class)
 public class MinecraftMixin implements MainThreadExecutor {
